feat: refactor common folder
