+++ src/views/Positions/Position.tsx
@@ -1,6 +1,6 @@
-import PositionName from 'src/views/Common/PositionName'
-import ProtocolIcon from 'src/views/Common/ProtocolIcon'
-import Title from 'src/views/Common/Title'
+import PositionName from 'src/views/Positions/PositionName'
+import ProtocolIcon from 'src/views/Positions/ProtocolIcon'
+import Title from 'src/views/Positions/Title'
 import BoxWrapperColumn from 'src/components/Wrappers/BoxWrapperColumn'
 import BoxWrapperRow from 'src/components/Wrappers/BoxWrapperRow'
 import * as React from 'react'
